Coffee Machine
This is a simple application programmed to create a request for a type of coffee. It includes the ingredients and price of the coffee. If the ingredients are insufficient, then the coffee request cannot be processed, and if the money is not enough to pay for the coffee, then the coffee cannot be processed.